+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Change Log
 All notable changes to the "vscode-postgres" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
+
+## [0.0.6] - 2018-04-30
+### Fixed
+- Repository link
+- Output Console Name for Language Service
+
+### Added
+- Ability to save results to JSON, XML, or CSV.
+
+## [0.0.5] - 2018-04-30
 - Initial release